first commit for fb-login
Attempting another setup, guide used axios so might have to add that.  For now, made the FB SDK into it's own file, along with react-router-dom history.  init-fb-sdk is called before the app starts so that might help with login.  Will need to fix the loginForm to include FB login button